Split api examples module into two
Split the Shelley genesis example from the script examples.

They are mostly unrelated and use very different imports. We will
shortly rewrite the script examples.
